move takeOrder to Vault. finish #10
