+++ lib/main.js
@@ -40,8 +40,6 @@ export sortByPrice from "./exchange/queries/sortByPrice";
 // ./lib/exchange/transactions
 export cancelOrder from "./exchange/transactions/cancelOrder";
 export makeOrder from "./exchange/transactions/makeOrder";
-export takeMultipleOrders from "./exchange/transactions/takeMultipleOrders";
-export takeOrder from "./exchange/transactions/takeOrder";
 
 // ./lib/exchange/utils
 export averagePrice from "./exchange/utils/averagePrice";
@@ -72,3 +70,5 @@ export performCalculations from "./vault/calls/performCalculations";
 
 // ./lib/vault/transactions
 export createVault from "./vault/transactions/createVault";
+export takeMultipleOrders from "./vault/transactions/takeMultipleOrders";
+export takeOrder from "./vault/transactions/takeOrder";
