+++ lib/vault/transactions/takeOrder.js
@@ -6,7 +6,7 @@ import VaultJson from "@melonproject/protocol/build/contracts/Vault.json";
 import setup from "../../utils/setup";
 import toProcessable from "../../assets/utils/toProcessable";
 import getConfig from "../../universe/calls/getConfig";
-import getOrder from "../calls/getOrder";
+import getOrder from "../../exchange/calls/getOrder";
 
 /*
   @param quantityAsked: BigNumber with Precision (i.e. '1.234' NOT '1234')
