+++ tests/unit/vault/transactions/takeMultipleOrders.js
@@ -2,7 +2,7 @@ import BigNumber from "bignumber.js";
 
 import matchedOrders from "../../../fixtures/matchedOrders";
 
-import takeMultipleOrders from "../../../../lib/exchange/transactions/takeMultipleOrders";
+import takeMultipleOrders from "../../../../lib/vault/transactions/takeMultipleOrders";
 
 /* eslint-disable global-require */
 jest.mock("truffle-contract", () => require("../../../mocks/truffle-contract"));
